# rope-manager-backend/openapi.yaml
openapi: 3.0.3
info:
  title: 绳包管理器后端 API
  description: |
    绳包管理器后端服务的完整API文档
    
    ## 认证方式
    系统支持两种认证方式，按优先级顺序：
    1. **Authorization 头部认证（推荐）**: `Authorization: Bearer <token>`
    2. **Cookie 认证**: `Cookie: auth_token=<token>`
    
    ## 权限等级
    - **普通用户 (user)**: 基础功能使用权限
    - **元老 (elder)**: 拥有部分管理权限，可审核资源
    - **版主 (moderator)**: 拥有内容管理权限
    - **管理员 (admin)**: 拥有所有系统权限
    
    ## API 响应格式
    所有API统一返回格式：
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {}
    }
    ```
    - code: 状态码，0表示成功，其他表示错误
    - message: 响应消息
    - data: 响应数据（可选）
    
  version: 1.0.0
  contact:
    name: 绳包管理器API支持
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:15201/api/v1
    description: 开发服务器 - API端点
  - url: http://127.0.0.1:15201
    description: 开发服务器 - 健康检查端点
security:
  - BearerAuth: []
  - CookieAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码，0表示成功
        message:
          type: string
          description: 响应消息
        data:
          description: 响应数据（可选）
      required:
        - code
        - message

    # 用户相关模型
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        author_id:
          type: integer
        author_name:
          type: string
          nullable: true
        category_id:
          type: integer
          nullable: true
        status:
          type: string
          enum: [Draft, Published, Archived, Deleted]
        view_count:
          type: integer
        like_count:
          type: integer
        comment_count:
          type: integer
        is_pinned:
          type: boolean
        is_featured:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category_id:
          type: integer
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [Draft, Published, Archived, Deleted]
      required:
        - title
        - content
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category_id:
          type: integer
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [Draft, Published, Archived, Deleted]
        is_pinned:
          type: boolean
        is_featured:
          type: boolean
    PostListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    # 标签相关模型
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        use_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /posts:
    get:
      tags:
        - 帖子系统
      summary: 获取帖子列表
      description: 获取帖子列表，支持分页和筛选
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
        - name: category_id
          in: query
          schema:
            type: integer
        - name: author_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [Draft, Published, Archived, Deleted]
        - name: search
          in: query
          schema:
            type: string
        - name: is_pinned
          in: query
          schema:
            type: boolean
        - name: is_featured
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PostListResponse'

    post:
      tags:
        - 帖子系统
      summary: 创建帖子
      description: 创建新帖子
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          post_id:
                            type: integer
        '401':
          description: 未授权
  /posts/featured:
    get:
      tags:
        - 帖子系统
      summary: 获取精华帖子
      description: 获取精华帖子列表
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PostListResponse'
  /posts/popular:
    get:
      tags:
        - 帖子系统
      summary: 获取热门帖子
      description: 获取热门帖子列表
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PostListResponse'
  /posts/{id}:
    get:
      tags:
        - 帖子系统
      summary: 获取帖子详情
      description: 根据ID获取帖子详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Post'
        '404':
          description: 帖子不存在

    put:
      tags:
        - 帖子系统
      summary: 更新帖子
      description: 更新帖子信息（作者或管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: 更新成功
        '403':
          description: 权限不足
        '404':
          description: 帖子不存在

    delete:
      tags:
        - 帖子系统
      summary: 删除帖子
      description: 删除帖子（作者或管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '403':
          description: 权限不足
        '404':
          description: 帖子不存在
  /posts/{id}/tags:
    get:
      tags:
        - 帖子系统
      summary: 获取帖子标签
      description: 获取指定帖子的标签列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tag'
  /posts/{id}/view:
    post:
      tags:
        - 帖子系统
      summary: 增加浏览量
      description: 增加帖子浏览量
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 操作成功
        '404':
          description: 帖子不存在

  # 标签管理接口
...